up node, cloud

+++ home.sh
@@ -27,6 +27,9 @@ read -r -p "Instalar o acesso remoto AnyDesk? [y|n] " ANYDESK
 read -r -p "Instalar o VSCode? [y|n] " VSCODE
 read -r -p "Instalar o VSCodium? [y|n] " VSCODIUM
 read -r -p "Instalar o Node JS? [y|n] " NODEJS
+if [ "$NODEJS" = "y" ]; then
+    read -r -p "Qual Versão do Node JS? [16|18|20] " NODEVERSION
+fi
 read -r -p "Instalar o SQLite? [y|n] " SQLITE
 read -r -p "Instalar o Antares SQL? [y|n] " ANTARESSQL
 read -r -p "Instalar o DWService? [y|n] " DWSERVICE
@@ -95,16 +98,22 @@ if [ "$CHROMIUM" = "y" ]; then
 fi
 #DropBox
 if [ "$DROPBOX" = "y" ]; then
-    nala install -y thunar-dropbox-plugin
+    #XFCE
+    if [ $XDG_CURRENT_DESKTOP = XFCE ]; then
+        nala install -y thunar-dropbox-plugin
+    fi
 fi
 #MegaSync
 if [ "$MEGASYNC" = "y" ]; then
     wget -c https://mega.nz/linux/repo/Debian_"$RELEASE"/amd64/megasync-Debian_"$RELEASE"_amd64.deb
-    wget -c https://mega.nz/linux/repo/Debian_"$RELEASE"/amd64/thunar-megasync-Debian_"$RELEASE"_amd64.deb
     nala install -y ./megasync*.deb
-    nala install -y ./thunar*.deb
     rm -R ./megasync*.deb
-    rm -R ./thunar*.deb
+    #XFCE
+    if [ $XDG_CURRENT_DESKTOP = XFCE ]; then
+        wget -c https://mega.nz/linux/repo/Debian_"$RELEASE"/amd64/thunar-megasync-Debian_"$RELEASE"_amd64.deb
+        nala install -y ./thunar*.deb
+        rm -R ./thunar*.deb
+    fi
 fi
 #NextCloud
 if [ "$NEXTCLOUD" = "y" ]; then
@@ -140,8 +149,7 @@ fi
 #Node JS
 if [ "$NODEJS" = "y" ]; then
     wget -qO - https://deb.nodesource.com/gpgkey/nodesource-repo.gpg.key | gpg --dearmor -o /etc/apt/keyrings/nodesource.gpg
-    NODE_MAJOR=20
-    echo "deb [signed-by=/etc/apt/keyrings/nodesource.gpg] https://deb.nodesource.com/node_$NODE_MAJOR.x nodistro main" | tee /etc/apt/sources.list.d/nodesource.list
+    echo "deb [signed-by=/etc/apt/keyrings/nodesource.gpg] https://deb.nodesource.com/node_$NODEVERSION.x nodistro main" | tee /etc/apt/sources.list.d/nodesource.list
     nala update
     nala install -y nodejs
 fi
@@ -168,6 +176,7 @@ if [ "$SOFTWAREDEB" = "y" ]; then
     nala install -y gnome-software
 else
     nala install -y deepin-deb-installer
+    nala install -y gnome-package-updater
 fi
 #Limpeza no apt
 nala autoremove -y
